[mod] LossRateDto 수정
