+++ .gitignore
@@ -45,3 +45,6 @@ nbe2_2_dev.trace.db
 
 #secret yml#
 application-secret.yml
+
+# Ignore ProductService file
+/src/main/kotlin/edu/example/kotlindevelop/domain/product/service/ProductService.kt
