+++ src/main/kotlin/edu/example/kotlindevelop/domain/product/dto/ProductDTO.kt
@@ -1,7 +1,8 @@
-package com.example.devcoursed.domain.product.product.dto
+package edu.example.devcoursed.domain.product.product.dto
 
 import edu.example.kotlindevelop.domain.product.entity.LossRate
 import edu.example.kotlindevelop.domain.member.entity.Member
+import edu.example.kotlindevelop.domain.product.dto.LossRateDTO
 import edu.example.kotlindevelop.domain.product.entity.Product
 import jakarta.validation.constraints.NotBlank
 import org.springframework.data.domain.PageRequest
@@ -10,10 +11,7 @@ import org.springframework.data.domain.Sort
 import java.math.BigDecimal
 import java.time.LocalDate
 
-class ProductDTO(
-    product: Product,
-    lossRate: LossRate
-) {
+class ProductDTO {
 
     data class CreateProductRequestDto(
         val id: Long,
@@ -21,7 +19,7 @@ class ProductDTO(
         @field:NotBlank(message = "식재료 이름은 필수 값입니다.")
         val name: String,
 
-        val lossRates: MutableList<LossRateRequestDTO> = mutableListOf()
+        val lossRates: MutableList<LossRateDTO.LossRateRequestDTO> = mutableListOf()
     ) {
         fun toEntity(member: Member): Product {
             return Product(
@@ -30,63 +28,6 @@ class ProductDTO(
             )
         }
     }
-
-    data class LossRateRequestDTO(
-
-        val id: Long? = null,
-
-        @field:NotBlank
-        val loss: Int,
-
-        val recordedAt: LocalDate? = null
-    ){
-        constructor(product: Product, lossRate: LossRate) : this(
-            id = product.id,
-            loss = lossRate.loss,
-            recordedAt = lossRate.recordedAt
-        )
-
-        fun toEntity(product: Product, member: Member): LossRate {
-            return LossRate(
-                maker = member,
-                product = product,
-                loss = this.loss
-            )
-        }
-
-    }
-
-    data class LossRateResponseDTO(
-        val productId: Long?,
-        val productName: String,
-        val lossRate: Int,
-        val recordedAt: LocalDate?
-    ) {
-        constructor(product: Product, lossRate: LossRate) : this(
-            productId = product.id,
-            productName = product.name,
-            lossRate = lossRate.loss,
-            recordedAt = lossRate.recordedAt
-        )
-    }
-
-
-
-
-
-//    constructor(product: Product) : this(
-//        id = product.id,
-//        name = product.name,
-//        loss = product.loss
-//    )
-//
-//    fun toEntity(member: Member?): Product {
-//        return Product(
-//            name = name,
-//            loss = loss,
-//            maker = member
-//        )
-//    }
 
     data class PageRequestDTO(
         private val page: Int = 0,
@@ -105,9 +46,4 @@ class ProductDTO(
             }
     }
 
-    data class AverageResponseDTO(
-        val dates: List<LocalDate>? = null,
-        val personalAverage: List<BigDecimal>? = null,
-        val allUsersAverage: List<BigDecimal>? = null
-    )
 }
